Move selection map to a file
